+++ src/cpu/instruction_set/increments_decrements/tests.rs
@@ -216,11 +216,11 @@ fn test_dey_negative_flag() {
         CpuFlags::NEGATIVE);
 }
 
-fn test_increment_decrement<F, G, H>(initial_value: u8, address: u8, expected_value: u8, set_value: F, get_value: G, operation: H, expected_status: CpuFlags)
+fn test_increment_decrement<F, G, H>(initial_value: u8, address: u16, expected_value: u8, set_value: F, get_value: G, operation: H, expected_status: CpuFlags)
 where
-    F: Fn(&mut Cpu, u8, u8),
-    G: Fn(&Cpu, u8) -> u8,
-    H: Fn(&mut Cpu, u8),
+    F: Fn(&mut Cpu, u16, u8),
+    G: Fn(&Cpu, u16) -> u8,
+    H: Fn(&mut Cpu, u16),
 {
     // Given
     let mut cpu = Cpu::new();
